requirements.dev.txt
boto3==1.28.49
botocore==1.31.49
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
tomli==2.0.1
httpcore==0.18.0
httpx==0.25.0
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.7.1
pydantic-core==2.18.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
urllib3<2
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
mangum==0.17.0
pyyaml==6.0.1
email_validator==2.0.0.post2
phonenumbers==8.13.22
python-jose[cryptography]
bcrypt==3.2.2
pre-commit==3.5.0
geopy
-r requirements.txt
